bot/requirements.txt
aiogram==3.3.0
sqlalchemy==2.0.27
pymysql==1.1.0
redis==5.0.1
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0 